Fixes and cleanups
* fluorescence.anisotropy renamed variables to be more specific
* gui.widgets.experiments.tcspc.controller use of pathlib
* gui.widgets.general use of pathlib
* gui.__init__.gui_imports() fixed broken import
* gui.main use of _current_dataset instead of access to PyQt widget
* macros.__init__ fixed typo
* settings use of pathlib
* settings update locals() from yaml dict
* specified type in parameter for

+++ chisurf/parameter.py
@@ -76,7 +76,7 @@ class Parameter(
             self,
             value: float
     ):
-        self._port.value = np.array([value])
+        self._port.value = np.array([value], dtype=np.double)
 
     @property
     def link(self) -> chisurf.parameter.Parameter:
